Front end views, controllers and models (#12)
* Added controllers and views for patients and beds

* Updated PatientApiController for some reusability

* Updated date time formatting and added _ViewStart for patients

* Add login requirement for actions, add admit patient to bed, add view bed details

* Fix errors not being displayed when admitting to bed, add a more informative error

* Update patient repository to clear health details on discharge correctly, add discharge front end routes and views, update repository tests

* Fix logout not working on stale logins, update bed list view

* Add get all employees list to interface, add tests

* Added employee creation

* Updated bed admission to track admissions per day

* Date time string update

* Connected comment functionality with views and controllers

* Fixed routing missing a route parameter

+++ Nulah.HospitalHelper.Lib/PatientManager.cs
@@ -195,7 +195,27 @@ namespace Nulah.HospitalHelper.Lib
                 return false;
             }
 
+            // Clear the patients presenting issue.
+            // This should return null on success
+            var healthDetails = _patientRepository.ClearHealthDetails(patientURN);
+
+            // Any non-null value indicates a failure
+            if (healthDetails != null)
+            {
+                return false;
+            }
+
             return _bedRepository.RemovePatientFromBed(patientURN, bedNumber);
+        }
+
+        /// <summary>
+        /// Returns the number of patients admitted for the given date
+        /// </summary>
+        /// <param name="date">Will be converted to UTC</param>
+        /// <returns></returns>
+        public int GetAdmittanceCountForDate(DateTime date)
+        {
+            return _patientRepository.GetAdmittanceStats(date.ToUniversalTime().Date);
         }
     }
 }
